// apps/fusion/project/ws/app/src/lib/routes/skill-management/models/competency-model.ts
export interface IApprover {
  skill_group: string
  role: string
  role_name: string
}

export interface ICourseObj {
  sequence_id: number
  course_type: string
  content_id: string
  content_name: string
}
export interface IAuthResponse {
  roles: IApprover[]
}
// ...
export interface ISkillList {
  category: string
  certification_count: number
  course_count: number
  criticality: string
  horizon: string
  skill_level?: string
  skill_level_id?: string
  image_url: string
  is_deleted: null
  is_grouped: null
  is_searchable: number
  rank_number: number
  skill_id: any
  skill_name: string
  skill_type: null
  skill_group?: string
  status?: string
}